+++ vnext/CHANGELOG.md
@@ -1,7 +1,37 @@
 # Change Log - react-native-windows
 
-This log was last generated on Tue, 10 Mar 2020 21:51:40 GMT and should not be manually modified.
+This log was last generated on Thu, 12 Mar 2020 19:18:01 GMT and should not be manually modified.
 
+## 0.61.0-beta.56
+Thu, 12 Mar 2020 19:18:01 GMT
+
+### Changes
+
+- Use redbox for native errors (acoates@microsoft.com)
+## 0.61.0-beta.55
+Thu, 12 Mar 2020 00:01:14 GMT
+
+### Changes
+
+- Fix JSValueObject operator[] (vmorozov@microsoft.com)
+## 0.61.0-beta.54
+Wed, 11 Mar 2020 19:32:20 GMT
+
+### Changes
+
+- Fully implement ExceptionsManagerModule (acoates@microsoft.com)
+## 0.61.0-beta.53
+Wed, 11 Mar 2020 18:14:12 GMT
+
+### Changes
+
+- Code review feedback (acoates@microsoft.com)
+## 0.61.0-beta.52
+Wed, 11 Mar 2020 16:18:00 GMT
+
+### Changes
+
+- Fix crash in timers during instance shutdown, and in systrace (acoates@microsoft.com)
 ## 0.61.0-beta.51
 Tue, 10 Mar 2020 21:51:40 GMT